test(governance): 拆分规范中心布局交互测试
